https://github.com/jackdewinter/pymarkdown/issues/594 - Refactoring link helper.
